requirements.txt
python-dotenv==1.0.0
neo4j==5.17.0
openai==1.6.0
streamlit==1.31.0